Start moving to a more conduity technique
I've had memory issues with large commits, that is, with a lot of changes, and
Attoparsec can't stream results, so I'm moving to Conduit so I don't need to
store as much in memory. I think it will allow for faster filtering, but it
does lose static guarantees, because you can make ridiculous streams (changes
that aren't in a commit, etc.).

I don't really know how to keep these while using Conduit. Particularly, I
would like to guarantee:

* Changes and `ls` without ref can only appear in a commit
* `data` always stream their output instead of parsing it, including commit
  messages

Another problem is when splitting branches, but I think when importing from SVN
usually a commit would only touch one branch, so optimizing for that, changes
belonging to the first branch would be emitted immediately, and changes to
other branches would be queued up. We'll see how that goes.