Add fancy conduit combinator for branch splitting
I'm not sure if I like it: It's slightly more general than it needs to be
(having only a subset that's allowed to stream immediately), and seems
to be doing three things. But I'm not sure how I would disentangle it.

Also, it's a bit long and unclear, and has a lot of lifts in it. I wonder
if it can be made more beautiful.